+++ lib/currency_formatter.dart
@@ -1,10 +1,12 @@
 library currency_formatter;
 
 import 'package:intl/intl.dart';
-import 'package:universal_io/io.dart' show Platform;
+import './src/locale.dart'
+    if (dart.library.html) './src/locale_html.dart'
+    if (dart.library.io) './src/locale_io.dart';
 
 abstract class CurrencyFormatter {
-  static final Map<num, String> _letters = {
+  static const Map<int, String> _letters = {
     1000000000000: 'T',
     1000000000: 'B',
     1000000: 'M',
@@ -25,7 +27,7 @@ abstract class CurrencyFormatter {
   /// e.g. `'$ 5.00'` instead of `'$ 5'`.
   static String format(
     amount,
-    CurrencyFormatterSettings settings, {
+    CurrencyFormat settings, {
     bool compact = false,
     int decimal = 2,
     showThousandSeparator = true,
@@ -44,22 +46,26 @@ abstract class CurrencyFormatter {
         }
       }
       number = amount.toStringAsPrecision(3);
-      number = number.replaceAll('.', settings.decimalSeparator!);
+      number = number.replaceAll('.', settings.decimalSeparator);
     } else {
       number = amount.toStringAsFixed(decimal);
-      if (!enforceDecimals && double.parse(number) == double.parse(number).round()) {
+      if (!enforceDecimals &&
+          double.parse(number) == double.parse(number).round()) {
         number = double.parse(number).round().toString();
       }
-      number = number.replaceAll('.', settings.decimalSeparator!);
+      number = number.replaceAll('.', settings.decimalSeparator);
       if (showThousandSeparator) {
-        String oldNum = number.split(settings.decimalSeparator!)[0];
-        number = number.contains(settings.decimalSeparator!)
-            ? settings.decimalSeparator! + number.split(settings.decimalSeparator!)[1]
+        String oldNum = number.split(settings.decimalSeparator)[0];
+        number = number.contains(settings.decimalSeparator)
+            ? settings.decimalSeparator +
+                number.split(settings.decimalSeparator)[1]
             : '';
         for (int i = 0; i < oldNum.length; i++) {
           number = oldNum[oldNum.length - i - 1] + number;
-          if ((i + 1) % 3 == 0 && i < oldNum.length - (oldNum.startsWith('-') ? 2 : 1))
-            number = settings.thousandSeparator! + number;
+          if ((i + 1) % 3 == 0 &&
+              i < oldNum.length - (oldNum.startsWith('-') ? 2 : 1)) {
+            number = settings.thousandSeparator + number;
+          }
         }
       }
     }
@@ -74,51 +80,51 @@ abstract class CurrencyFormatter {
   }
 
   /// Parse a formatted string to a number.
-  static num parse(String input, CurrencyFormatterSettings settings) {
+  static num parse(String input, CurrencyFormat settings) {
     String txt = input
-        .replaceFirst(settings.thousandSeparator ?? '', '')
-        .replaceFirst(settings.decimalSeparator ?? '', '.')
+        .replaceFirst(settings.thousandSeparator, '')
+        .replaceFirst(settings.decimalSeparator, '.')
         .replaceFirst(settings.symbol, '')
         .replaceFirst(settings.symbolSeparator, '')
         .trim();
 
-    String _letter = '';
-    for (String letter in _letters.values) {
+    int multiplicator = 1;
+    for (int mult in _letters.keys) {
+      final String letter = _letters[mult]!;
       if (txt.endsWith(letter)) {
         txt = txt.replaceFirst(letter, '');
-        _letter = letter;
+        multiplicator = mult;
         break;
       }
     }
 
-    return num.parse(txt) *
-        _letters.keys.firstWhere((e) => _letters[e] == _letter, orElse: () => 1);
+    return num.parse(txt) * multiplicator;
   }
 
-  /// Map that contains the [CurrencyFormatterSettings] from major currencies.
+  /// Map that contains the [CurrencyFormat] from major currencies.
   /// They can be accessed using their abbreviation. e.g. `majors['usd']`.
-  static final Map<String, CurrencyFormatterSettings> majors = {
-    'usd': CurrencyFormatterSettings.usd,
-    'eur': CurrencyFormatterSettings.eur,
-    'jpy': CurrencyFormatterSettings.jpy,
-    'gbp': CurrencyFormatterSettings.gbp,
-    'chf': CurrencyFormatterSettings.chf,
-    'cny': CurrencyFormatterSettings.cny,
-    'sek': CurrencyFormatterSettings.sek,
-    'krw': CurrencyFormatterSettings.krw,
-    'inr': CurrencyFormatterSettings.inr,
-    'rub': CurrencyFormatterSettings.rub,
-    'zar': CurrencyFormatterSettings.zar,
-    'try': CurrencyFormatterSettings.tryx,
-    'pln': CurrencyFormatterSettings.pln,
-    'thb': CurrencyFormatterSettings.thb,
-    'idr': CurrencyFormatterSettings.idr,
-    'huf': CurrencyFormatterSettings.huf,
-    'czk': CurrencyFormatterSettings.czk,
-    'ils': CurrencyFormatterSettings.ils,
-    'php': CurrencyFormatterSettings.php,
-    'myr': CurrencyFormatterSettings.myr,
-    'ron': CurrencyFormatterSettings.ron
+  static const Map<String, CurrencyFormat> majors = {
+    'usd': CurrencyFormat.usd,
+    'eur': CurrencyFormat.eur,
+    'jpy': CurrencyFormat.jpy,
+    'gbp': CurrencyFormat.gbp,
+    'chf': CurrencyFormat.chf,
+    'cny': CurrencyFormat.cny,
+    'sek': CurrencyFormat.sek,
+    'krw': CurrencyFormat.krw,
+    'inr': CurrencyFormat.inr,
+    'rub': CurrencyFormat.rub,
+    'zar': CurrencyFormat.zar,
+    'try': CurrencyFormat.tryx,
+    'pln': CurrencyFormat.pln,
+    'thb': CurrencyFormat.thb,
+    'idr': CurrencyFormat.idr,
+    'huf': CurrencyFormat.huf,
+    'czk': CurrencyFormat.czk,
+    'ils': CurrencyFormat.ils,
+    'php': CurrencyFormat.php,
+    'myr': CurrencyFormat.myr,
+    'ron': CurrencyFormat.ron
   };
 
   /// Map that contains the symbols from major currencies.
@@ -148,47 +154,54 @@ abstract class CurrencyFormatter {
   };
 }
 
+@Deprecated('Use \'CurrencyFormat\' instead')
+typedef CurrencyFormatterSettings = CurrencyFormat;
+
 /// This class contains the formatting settings for a currency.
-class CurrencyFormatterSettings {
+class CurrencyFormat {
   /// Symbol of the currency. e.g. '€'
-  String symbol;
+  final String symbol;
 
   /// Whether the symbol is shown before or after the money value, or if it is not shown at all.
   /// e.g. $ 125 ([SymbolSide.left]) or 125 € ([SymbolSide.right]).
-  SymbolSide symbolSide;
+  final SymbolSide symbolSide;
 
-  /// Thousand separator. e.g. 1,000,000 (`','`) or 1.000.000 (`'.'`). It can be set to any desired [String].
-  /// It defaults to `','` for [SymbolSide.left] and to `'.'` for [SymbolSide.right].
-  String? thousandSeparator;
+  final String? _thousandSeparator;
 
-  /// Decimal separator. e.g. 9.10 (`'.'`) or 9,10 (`','`). It can be set to any desired [String].
-  /// It defaults to `'.'` for [SymbolSide.left] and to `','` for [SymbolSide.right].
-  String? decimalSeparator;
+  final String? _decimalSeparator;
 
   /// Character(s) between the number and the currency symbol. e.g. $ 9.10 (`' '`) or $9.10 (`''`).
   /// It defaults to a normal space (`' '`).
-  String symbolSeparator;
+  final String symbolSeparator;
 
-  CurrencyFormatterSettings({
+  const CurrencyFormat({
     required this.symbol,
     this.symbolSide = SymbolSide.left,
-    this.thousandSeparator,
-    this.decimalSeparator,
+    String? thousandSeparator,
+    String? decimalSeparator,
     this.symbolSeparator = ' ',
-  }) {
-    thousandSeparator ??= symbolSide == SymbolSide.left ? ',' : '.';
-    decimalSeparator ??= symbolSide == SymbolSide.left ? '.' : ',';
-  }
+  })  : _thousandSeparator = thousandSeparator,
+        _decimalSeparator = decimalSeparator;
 
-  // Returns the same [CurrencyFormatterSettings] but with some changed parameters.
-  CurrencyFormatterSettings copyWith({
+  /// Thousand separator. e.g. 1,000,000 (`','`) or 1.000.000 (`'.'`). It can be set to any desired [String].
+  /// It defaults to `','` for [SymbolSide.left] and to `'.'` for [SymbolSide.right].
+  String get thousandSeparator =>
+      _thousandSeparator ?? (symbolSide == SymbolSide.left ? ',' : '.');
+
+  /// Decimal separator. e.g. 9.10 (`'.'`) or 9,10 (`','`). It can be set to any desired [String].
+  /// It defaults to `'.'` for [SymbolSide.left] and to `','` for [SymbolSide.right].
+  String get decimalSeparator =>
+      _decimalSeparator ?? (symbolSide == SymbolSide.left ? '.' : ',');
+
+  /// Returns the same [CurrencyFormat] but with some changed parameters.
+  CurrencyFormat copyWith({
     String? symbol,
     SymbolSide? symbolSide,
     String? thousandSeparator,
     String? decimalSeparator,
     String? symbolSeparator,
   }) =>
-      CurrencyFormatterSettings(
+      CurrencyFormat(
         symbol: symbol ?? this.symbol,
         symbolSide: symbolSide ?? this.symbolSide,
         thousandSeparator: thousandSeparator ?? this.thousandSeparator,
@@ -196,81 +209,124 @@ class CurrencyFormatterSettings {
         symbolSeparator: symbolSeparator ?? this.symbolSeparator,
       );
 
-  /// Get the [CurrencyFormatterSettings] of a currency using its symbol.
-  static CurrencyFormatterSettings? fromSymbol(String symbol) {
+  /// Get the [CurrencyFormat] of a currency using its symbol.
+  static CurrencyFormat? fromSymbol(String symbol) {
     for (int i = 0; i < CurrencyFormatter.majorSymbols.length; i++) {
-      if (CurrencyFormatter.majorSymbols.values.elementAt(i) == symbol)
+      if (CurrencyFormatter.majorSymbols.values.elementAt(i) == symbol) {
         return CurrencyFormatter.majors.values.elementAt(i);
+      }
     }
     return null;
   }
 
-  /// Get the [CurrencyFormatterSettings] of the local currency.
-  static CurrencyFormatterSettings? get local =>
-      fromSymbol(NumberFormat.simpleCurrency(locale: Platform.localeName).currencySymbol);
+  /// Get the [CurrencyFormat] of the local currency.
+  static CurrencyFormat? get local => fromSymbol(
+      NumberFormat.simpleCurrency(locale: localeName).currencySymbol);
 
-  static final CurrencyFormatterSettings usd =
-      CurrencyFormatterSettings(symbol: '\$', symbolSide: SymbolSide.left);
+  /// The [CurrencyFormat] of the US Dollar.
+  static const CurrencyFormat usd =
+      CurrencyFormat(symbol: '\$', symbolSide: SymbolSide.left);
 
-  static final CurrencyFormatterSettings eur =
-      CurrencyFormatterSettings(symbol: '€', symbolSide: SymbolSide.right);
+  /// The [CurrencyFormat] of the Euro.
+  static const CurrencyFormat eur =
+      CurrencyFormat(symbol: '€', symbolSide: SymbolSide.right);
 
-  static final CurrencyFormatterSettings jpy =
-      CurrencyFormatterSettings(symbol: '¥', symbolSide: SymbolSide.left);
+  /// The [CurrencyFormat] of the Japanese Yen.
+  static const CurrencyFormat jpy =
+      CurrencyFormat(symbol: '¥', symbolSide: SymbolSide.left);
 
-  static final CurrencyFormatterSettings gbp =
-      CurrencyFormatterSettings(symbol: '£', symbolSide: SymbolSide.left);
+  /// The [CurrencyFormat] of the British Pound.
+  static const CurrencyFormat gbp =
+      CurrencyFormat(symbol: '£', symbolSide: SymbolSide.left);
 
-  static final CurrencyFormatterSettings chf =
-      CurrencyFormatterSettings(symbol: 'fr', symbolSide: SymbolSide.right);
+  /// The [CurrencyFormat] of the Swiss Franc.
+  static const CurrencyFormat chf =
+      CurrencyFormat(symbol: 'fr', symbolSide: SymbolSide.right);
 
-  static final CurrencyFormatterSettings cny =
-      CurrencyFormatterSettings(symbol: '元', symbolSide: SymbolSide.left);
+  /// The [CurrencyFormat] of the Chinese Yuan.
+  static const CurrencyFormat cny =
+      CurrencyFormat(symbol: '元', symbolSide: SymbolSide.left);
 
-  static final CurrencyFormatterSettings sek =
-      CurrencyFormatterSettings(symbol: 'kr', symbolSide: SymbolSide.right);
+  /// The [CurrencyFormat] of the Swedish Krona.
+  static const CurrencyFormat sek =
+      CurrencyFormat(symbol: 'kr', symbolSide: SymbolSide.right);
 
-  static final CurrencyFormatterSettings krw =
-      CurrencyFormatterSettings(symbol: '₩', symbolSide: SymbolSide.left);
+  /// The [CurrencyFormat] of the South Korean Won.
+  static const CurrencyFormat krw =
+      CurrencyFormat(symbol: '₩', symbolSide: SymbolSide.left);
 
-  static final CurrencyFormatterSettings inr =
-      CurrencyFormatterSettings(symbol: '₹', symbolSide: SymbolSide.left);
+  /// The [CurrencyFormat] of the Indian Rupee.
+  static const CurrencyFormat inr =
+      CurrencyFormat(symbol: '₹', symbolSide: SymbolSide.left);
 
-  static final CurrencyFormatterSettings rub =
-      CurrencyFormatterSettings(symbol: '₽', symbolSide: SymbolSide.right);
+  /// The [CurrencyFormat] of the Russian Ruble.
+  static const CurrencyFormat rub =
+      CurrencyFormat(symbol: '₽', symbolSide: SymbolSide.right);
 
-  static final CurrencyFormatterSettings zar =
-      CurrencyFormatterSettings(symbol: 'R', symbolSide: SymbolSide.left);
+  /// The [CurrencyFormat] of the South African Rand.
+  static const CurrencyFormat zar =
+      CurrencyFormat(symbol: 'R', symbolSide: SymbolSide.left);
 
-  static final CurrencyFormatterSettings tryx =
-      CurrencyFormatterSettings(symbol: '₺', symbolSide: SymbolSide.left);
+  /// The [CurrencyFormat] of the Turkish Lira.
+  static const CurrencyFormat tryx =
+      CurrencyFormat(symbol: '₺', symbolSide: SymbolSide.left);
 
-  static final CurrencyFormatterSettings pln =
-      CurrencyFormatterSettings(symbol: 'zł', symbolSide: SymbolSide.right);
+  /// The [CurrencyFormat] of the Polish Zloty.
+  static const CurrencyFormat pln =
+      CurrencyFormat(symbol: 'zł', symbolSide: SymbolSide.right);
 
-  static final CurrencyFormatterSettings thb =
-      CurrencyFormatterSettings(symbol: '฿', symbolSide: SymbolSide.left);
+  /// The [CurrencyFormat] of the Thai Baht.
+  static const CurrencyFormat thb =
+      CurrencyFormat(symbol: '฿', symbolSide: SymbolSide.left);
 
-  static final CurrencyFormatterSettings idr =
-      CurrencyFormatterSettings(symbol: 'Rp', symbolSide: SymbolSide.left);
+  /// The [CurrencyFormat] of the Indonesian Rupiah.
+  static const CurrencyFormat idr =
+      CurrencyFormat(symbol: 'Rp', symbolSide: SymbolSide.left);
 
-  static final CurrencyFormatterSettings huf =
-      CurrencyFormatterSettings(symbol: 'Ft', symbolSide: SymbolSide.right);
+  /// The [CurrencyFormat] of the Hungarian Forint.
+  static const CurrencyFormat huf =
+      CurrencyFormat(symbol: 'Ft', symbolSide: SymbolSide.right);
 
-  static final CurrencyFormatterSettings czk =
-      CurrencyFormatterSettings(symbol: 'Kč', symbolSide: SymbolSide.right);
+  /// The [CurrencyFormat] of the Czech Koruna.
+  static const CurrencyFormat czk =
+      CurrencyFormat(symbol: 'Kč', symbolSide: SymbolSide.right);
 
-  static final CurrencyFormatterSettings ils =
-      CurrencyFormatterSettings(symbol: '₪', symbolSide: SymbolSide.left);
+  /// The [CurrencyFormat] of the Israeli New Shekel.
+  static const CurrencyFormat ils =
+      CurrencyFormat(symbol: '₪', symbolSide: SymbolSide.left);
 
-  static final CurrencyFormatterSettings php =
-      CurrencyFormatterSettings(symbol: '₱', symbolSide: SymbolSide.left);
+  /// The [CurrencyFormat] of the Philippine Peso.
+  static const CurrencyFormat php =
+      CurrencyFormat(symbol: '₱', symbolSide: SymbolSide.left);
 
-  static final CurrencyFormatterSettings myr =
-      CurrencyFormatterSettings(symbol: 'RM', symbolSide: SymbolSide.left);
+  /// The [CurrencyFormat] of the Malaysian Ringgit.
+  static const CurrencyFormat myr =
+      CurrencyFormat(symbol: 'RM', symbolSide: SymbolSide.left);
 
-  static final CurrencyFormatterSettings ron =
-      CurrencyFormatterSettings(symbol: 'L', symbolSide: SymbolSide.right);
+  /// The [CurrencyFormat] of the Romanian Leu.
+  static const CurrencyFormat ron =
+      CurrencyFormat(symbol: 'L', symbolSide: SymbolSide.right);
+
+  @override
+  String toString() =>
+      'CurrencyFormat<${CurrencyFormatter.format(9999.99, this)}>';
+
+  @override
+  operator ==(other) =>
+      other is CurrencyFormat &&
+      other.symbol == symbol &&
+      other.symbolSide == symbolSide &&
+      other.thousandSeparator == thousandSeparator &&
+      other.decimalSeparator == decimalSeparator &&
+      other.symbolSeparator == symbolSeparator;
+
+  @override
+  int get hashCode =>
+      symbol.hashCode ^
+      symbolSide.hashCode ^
+      thousandSeparator.hashCode ^
+      decimalSeparator.hashCode ^
+      symbolSeparator.hashCode;
 }
 
 /// Enumeration for the three possibilities when writing the currency symbol.